Closed #15. Added graph into another graph. CleanUp Code
