+++ app/src/main/java/com/example/composeapplication/data/BottomNavigationItem.kt
@@ -1,10 +1,14 @@
 package com.example.composeapplication.data
 
-import com.example.composeapplication.R
+import androidx.compose.material.icons.Icons
+import androidx.compose.material.icons.filled.AccountBox
+import androidx.compose.material.icons.filled.Info
+import androidx.compose.material.icons.filled.Person
+import androidx.compose.ui.graphics.vector.ImageVector
 
-sealed class BottomNavigationItem(var title: String, var icon: Int, var screen_route: String) {
+sealed class BottomNavigationItem(var title: String, var icon: ImageVector, var screen_route: String) {
 
-    object Profile : BottomNavigationItem("Profile", R.drawable.icon_profile, "profile")
-    object ProfileList : BottomNavigationItem("ProfileList", R.drawable.icon_list, "profileList")
-    object RecipeList : BottomNavigationItem("RecipeList", R.drawable.icon_list, "RecipeList")
+    object Profile : BottomNavigationItem("Profile", Icons.Default.Person, "profile")
+    object ProfileList : BottomNavigationItem("ProfileList", Icons.Default.AccountBox, "profileList")
+    object RecipeList : BottomNavigationItem("RecipeList", Icons.Default.Info, "RecipeList")
 }
