+++ app/src/main/java/com/example/composeapplication/view/common/BottomBarNavigation.kt
@@ -1,9 +1,16 @@
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.layout.wrapContentHeight
 import androidx.compose.material3.BottomAppBar
 import androidx.compose.material3.Icon
+import androidx.compose.material3.MaterialTheme
 import androidx.compose.material3.NavigationBarItem
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.getValue
+import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.painterResource
+import androidx.compose.ui.unit.dp
 import androidx.navigation.NavHostController
 import androidx.navigation.compose.currentBackStackEntryAsState
 import com.example.composeapplication.data.BottomNavigationItem
@@ -15,7 +22,9 @@ fun BottomBar(navController: NavHostController) {
         BottomNavigationItem.ProfileList,
     )
 
-    BottomAppBar {
+    BottomAppBar(modifier = Modifier
+        .wrapContentHeight()
+        .fillMaxWidth()) {
         val navBackStackEntry by navController.currentBackStackEntryAsState()
         val currentRoute = navBackStackEntry?.destination?.route
         screens.forEach { item ->
@@ -32,7 +41,16 @@ fun BottomBar(navController: NavHostController) {
                     }
                 },
                 alwaysShowLabel = true,
-                icon = { Icon(painterResource(id = item.icon), contentDescription = item.title) })
+                icon = {
+                    Icon(
+                        modifier= Modifier
+                            .width(32.dp)
+                            .padding(8.dp),
+                        tint=MaterialTheme.colorScheme.onBackground,
+                        imageVector = item.icon,
+                        contentDescription = item.title)
+                }
+            )
         }
 
     }
